requirements.txt
appdirs==1.4.4
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2
django-allauth==0.44.0
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
filelock==3.0.12
gunicorn==20.1.0
idna==2.10
oauthlib==3.1.1
Pillow==8.2.0
psycopg2==2.8.6
pycparser==2.20
pycryptodome==3.10.1
PyJWT==2.1.0
pypng==0.0.20
PyQRCode==1.2.1
python-decouple==3.4
python-dotenv==0.18.0
python-environ==0.4.54
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.5
validate-email==1.3
virtualenv==20.4.4
virtualenvwrapper-win==1.2.6
whitenoise==5.2.0
